Вынес реализацию сервисов из контроллера в конкст БД
